BIG ONE INCOMING
Okay so we fixed the chat screen and got it working by simplifying the localstorage code. Also hooked up the ai function to npms Ollama package which is working flawlessly. tinkered with a few different llms but am pretty certain ill need to find one that more geared to longform story telling. Also created a sick spinner for when the ai is processing the data.